Add docstring and move tokenizer
